fix a crash with conversations
When someone sends me a message, the irc::Buddy object is pointed by
im::Conversation, but irc::Buddy hasn't his 'conv' attribute set. So if I
don't send any message, and if the buddy disconnects from IRC (so when the
account is disabled), the irc::Buddy's destructor will not reset the
pointer to him in the im::Conversation object.