Elements of RPL_MYINFO reply should be separate IRC-protocol arguments.
Before this patch, minbif would send a RPL_MYINFO numeric that looked like:

:harzilein.eu.org 004 Harzilein :harzilein.eu.org minbif-1.0.5(caca)(pam)(tls) o qohvb

This is parsed by clients as two arguments: "Harzilein" and
"harzilein.eu.org minbif-1.0.5(caca)(pam)(tls) o qohvb".  This is incorrect,
and causes at least EPIC4 to fail to recognise the 004 numeric.

After this patch, it is now sent as separate arguments.

Signed-off-by: Kevin Easton <caf@guarana.org>